AdcMatrix block-wise apply tests
